requirements.txt
streamlit
langchain
langchain-openai
sqlite3
